Trait serialization::Serializable

pub trait Serializable {
    fn serialize(&self, s: &mut Stream);

    fn serialized_size(&self) -> usize
    where
        Self: Sized,
    { ... }

    fn serialized_size_with_flags(&self, flags: u32) -> usize
    where
        Self: Sized,
    { ... }
}

Required Methods

fn serialize(&self, s: &mut Stream)

Serialize the struct and appends it to the end of stream.

Provided Methods

fn serialized_size(&self) -> usize where
Self: Sized,

Hint about the size of serialized struct.

fn serialized_size_with_flags(&self, flags: u32) -> usize where
Self: Sized,

Hint about the size of serialized struct with given flags.